#57/#162 - Admin View Updates + Backend (server.js) code refactoring
