+++ GlamByManpreet/package.json
@@ -1,7 +1,6 @@
 {
   "name": "my-react-app",
   "version": "0.1.0",
-  
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.13.3",
